LayoutTests/editing/spelling/spelling-double-clicked-word.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="resources/util.js"></script>
</head>
<body>
<div id="container">
  <div id="destination" contentEditable></div>
</div>

<script>

description("Spelling should work for double-clicked misspellings. " +
            "To test manually, type 'wellcome home.', double-click on 'wellcome', and context-click on the selection. " +
            "The test succeeds when the context menu shows spell check suggestions.");

initSpellTest("destination", "wellcome home.", function(textNode) {
    spellingMarkerRange = internals.markerRangeForNode(textNode, "spelling", 0);
    shouldBeEqualToString("spellingMarkerRange.toString()", "wellcome");
    shouldBeEqualToString("window.getSelection().toString()", "");

    // Double-click the misspelled word without trailing whitespace.
    testRunner.setSelectTrailingWhitespaceEnabled(false);
    var x = destination.offsetParent.offsetLeft + destination.offsetLeft + 40;
    var y = destination.offsetParent.offsetTop + destination.offsetTop + destination.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
    eventSender.mouseDown();
    eventSender.mouseUp();
    eventSender.mouseDown();
    eventSender.mouseUp();

    shouldBeEqualToString("window.getSelection().toString()", "wellcome");

    // Context click on "wellcome" to show the context menu.
    x = destination.offsetParent.offsetLeft + destination.offsetLeft + 40;
    y = destination.offsetParent.offsetTop + destination.offsetTop + destination.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
    contextMenuElements = eventSender.contextClick();
    // Esc key to hide the context menu.
    eventSender.keyDown(String.fromCharCode(0x001B), null);

    spellingMarkerRange = internals.markerRangeForNode(textNode, "spelling", 0);
    shouldBeEqualToString("spellingMarkerRange.toString()", "wellcome");
    shouldBeEqualToString("window.getSelection().toString()", "wellcome");
    shouldBeEqualToString("contextMenuElements[contextMenuElements.length - 1]", "welcome");

    window.getSelection().removeAllRanges();
    shouldBeEqualToString("window.getSelection().toString()", "");

    // Double-click the misspelled word with trailing whitespace.
    testRunner.setSelectTrailingWhitespaceEnabled(true);
    var x = destination.offsetParent.offsetLeft + destination.offsetLeft + 40;
    var y = destination.offsetParent.offsetTop + destination.offsetTop + destination.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
    eventSender.mouseDown();
    eventSender.mouseUp();
    eventSender.mouseDown();
    eventSender.mouseUp();
    shouldBeEqualToString("window.getSelection().toString()", "wellcome ");

    // Context click on "wellcome " to show the context menu.
    x = destination.offsetParent.offsetLeft + destination.offsetLeft + 40;
    y = destination.offsetParent.offsetTop + destination.offsetTop + destination.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
    contextMenuElements = eventSender.contextClick();
    // Esc key to hide the context menu.
    eventSender.keyDown(String.fromCharCode(0x001B), null);

    spellingMarkerRange = internals.markerRangeForNode(textNode, "spelling", 0);
    shouldBeEqualToString("spellingMarkerRange.toString()", "wellcome");
    shouldBeEqualToString("window.getSelection().toString()", "wellcome ");
    shouldBeEqualToString("contextMenuElements[contextMenuElements.length - 1]", "welcome");
});

</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
