LayoutTests/editing/spelling/spelling-double-clicked-word-expected.txt

Spelling should work for double-clicked misspellings. To test manually, type 'wellcome home.', double-click on 'wellcome', and context-click on the selection. The test succeeds when the context menu shows spell check suggestions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is ""
PASS window.getSelection().toString() is "wellcome"
PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is "wellcome"
PASS contextMenuElements[contextMenuElements.length - 1] is "welcome"
PASS window.getSelection().toString() is ""
PASS window.getSelection().toString() is "wellcome "
PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is "wellcome "
PASS contextMenuElements[contextMenuElements.length - 1] is "welcome"
PASS successfullyParsed is true

TEST COMPLETE
wellcome home.
